test(licensing): pin the write's WHERE structure, not just its fragments
The re-review found three mutations the previous assertions could not see, all in
the same blind spot: the tests read the shared predicate fragment and the values
bound to each statement, and nothing read the SQL that JOINS them.

- `WHERE m.id = ANY(...) OR <predicate>` passed every assertion. In production
  that statement rewrites the whole population in one go while the response and
  the enqueue name one batch. The write's WHERE shape is now asserted whole, with
  whitespace collapsed so a re-indent is not a false red.
- A second write statement was invisible, because every write assertion reads
  call 0. Re-adding the sibling's updatedAt bump as its own statement, or through
  $executeRaw, passed. The test now pins one $queryRaw call and no $executeRaw.
- The keyset test asserted the bound values by membership, so swapping the
  afterId and batchSize interpolations passed — the endpoint would page from
  id > 1200 in batches of 4000. The numbers are now asserted in order.

Also from the re-review: predicatesOf returns a list and the test asserts exactly
one fragment per statement before comparing identity, because two undefineds are
toBe each other; batchSize takes .int(), since a fractional LIMIT is rounded by
the Prisma engine and rejected by node-postgres; and the header no longer claims
every one of the 153 rows was saved through a build offering the checkbox — the
ones nearest the boundary may predate the roll, which is an under-grant and the
safe direction. The asymmetry that justifies the whole cutoff choice is restored
to the header, and the 153-row figure now carries its provenance.

Six mutation controls re-run: OR outside the fragment, a second statement, the
swapped interpolations, a hand-copied predicate, ::timestamptz, and the updatedAt
bump. Each reddened the test named for it.

One regression in the fix round, caught by eslint rather than by the suite: the
edit that rewrote the identity test dropped its `toContain(PREDICATE_BLOCK)` line,
leaving that constant unused and an OR INSIDE the fragment unguarded. Restored,
and its own control now reddens.

+++ src/pages/api/admin/temp/backfill-sell-merge-licence.ts
@@ -61,12 +61,24 @@ import { booleanString } from '~/utils/zod-helpers';
  * is not a decision and this may grant it. A row touched after that has been through a form offering
  * the box, so leaving it unticked IS a decision and must survive this run.
  *
+ * Erring early skips legacy rows; erring late overwrites a stated permission. That asymmetry is why
+ * this edge is the earliest defensible one rather than the latest.
+ *
  * `v5.1.103` (`a0cf2cceb6`, committed 2026-09-16 21:49:42Z) is the FIRST release containing #4874,
  * which opened the write paths — not `v5.1.105`, whose image timestamp (2026-09-17 03:11:25Z) was
- * originally taken for this edge and is about five hours later. Nothing can serve a build before its
- * commit exists, so the commit time is the conservative edge. Measured on the prod replica, the two
- * cutoffs differ by 153 rows, and every one of those is a creator's save through a build that offered
- * the checkbox.
+ * originally taken for this edge and is 5h21m later. Prod serves tagged releases, so the relevant
+ * commit is that tag's, not #4874's own merge commit (2026-09-16 19:10:43Z) — the merge is 2h39m
+ * earlier and no build existed at it. Nothing can serve a release before its commit exists, so the
+ * tag's commit time is the conservative edge, and it is re-derivable from `git log` by anyone without
+ * a database query.
+ *
+ * The two cutoffs differ by 153 rows (prod REPLICA, `DATABASE_REPLICA_URL` on 25061, measured
+ * 2026-09-17 from the primary checkout; 422,634 in scope at the old edge against 422,481 at this
+ * one). Most of those are saves through a build that offered the checkbox; the ones closest to the
+ * boundary may instead have been saved by v5.1.102 in the window between the tag and the roll, which
+ * did not offer it. Those are legacy rows this run now skips — an under-grant, which is the safe
+ * direction. The argument rests on the release ordering rather than on the count, so treat 153 as an
+ * illustration that moves with organic churn rather than as a constant.
  *
  * The evidence that build was SERVING rather than merely tagged: model 2943312, a Trained upload, was
  * born 2026-09-16 22:13:58.728Z carrying `{Image,RentCivit,Rent,Sell,SellMerge}` with
@@ -147,7 +159,9 @@ export function contractAcceptsSellMerge(
 
 const schema = z.object({
   dryRun: booleanString().default(true),
-  batchSize: z.coerce.number().min(1).max(5000),
+  // `.int()` because a fractional batchSize reaches Postgres as a fractional LIMIT, which the Prisma
+  // engine rounds (measured: `LIMIT 2.5` returns 3 rows) while node-postgres rejects it outright.
+  batchSize: z.coerce.number().int().min(1).max(5000),
   afterId: z.coerce.number().min(0).default(0),
 });
